// src/app/precisa/maestros/lista-base/model/listabase.ts
export class listabase{

    constructor(){
    }
    IdListaBase: number;
    Moneda: string;    
    TipoLista: string;
    IdCliente?: number;
    IndicadorPrecioFijo?: number;
    IdEmpresaAnteriorUnificacion?: number;
    TipoPaciente: number;
    FechaValidezInicio?: Date;
    FechaValidezFin?: Date;
    Codigo: string;
    Nombre: string;
    Descripcion: string;
    Estado: number;
    UsuarioCreacion: string;
    FechaCreacion: Date;
    UsuarioModificacion: string;
    FechaModificacion: Date;
    ESTADOdesc: string;

}